# perf: cleanup superfluous allocations during event unmarshal