## Which problem is this PR solving?
Waste in the existing unmarshal() implementation remaining from the following PR:
- #1586

## Short description of the changes
Skips the expensive Reader-based vmihailenco/msgpack.Decoder implementation for the common case of messagepack batches. Also de-pointerifies Payload's internal MsgpPayloadMap which was another heap allocation for no good reason. This is in no way comprehensive; there is other fat to trim in the batch ingest path.

```
benchmark                   old ns/op     new ns/op     delta
BenchmarkRouterBatch-12     5857          3724          -36.42%

benchmark                   old allocs     new allocs     delta
BenchmarkRouterBatch-12     69             44             -36.23%

benchmark                   old bytes     new bytes     delta
BenchmarkRouterBatch-12     5535          3550          -35.86%
```